fix len type of path parser

+++ gaba.h
@@ -345,7 +345,7 @@ int64_t gaba_dp_print_cigar(
 	FILE *fp,
 	uint32_t const *path,
 	uint32_t offset,
-	int64_t len);
+	uint32_t len);
 
 /**
  * @fn gaba_dp_dump_cigar
@@ -354,7 +354,7 @@ int64_t gaba_dp_dump_cigar(
 	char *buf,
 	uint32_t const *path,
 	uint32_t offset,
-	int64_t len);
+	uint32_t len);
 
 #endif  /* #ifndef _GABA_H_INCLUDED */
